+++ gulpfile.ts
@@ -1,13 +1,15 @@
 import * as buildTools from "turbo-gulp";
+import { LibTarget, registerLibTasks } from "turbo-gulp/targets/lib";
+import { MochaTarget, registerMochaTasks } from "turbo-gulp/targets/mocha";
 
 import gulp from "gulp";
-import minimist, { ParsedArgs } from "minimist";
+import minimist from "minimist";
 
 interface Options {
   devDist?: string;
 }
 
-const options: Options & ParsedArgs = minimist(process.argv.slice(2), {
+const options: Options & minimist.ParsedArgs = minimist(process.argv.slice(2), {
   string: ["devDist"],
   default: {devDist: undefined},
   alias: {devDist: "dev-dist"},
@@ -21,13 +23,12 @@ const project: buildTools.Project = {
   srcDir: "src",
 };
 
-const lib: buildTools.LibTarget = {
+const lib: LibTarget = {
   project,
   name: "lib",
   srcDir: "src/lib",
   scripts: ["**/*.ts"],
   mainModule: "index",
-  outModules: buildTools.OutModules.Js,
   dist: {
     packageJsonMap: (old: buildTools.PackageJson): buildTools.PackageJson => {
       const version: string = options.devDist !== undefined ? `${old.version}-build.${options.devDist}` : old.version;
@@ -38,7 +39,8 @@ const lib: buildTools.LibTarget = {
     },
   },
   customTypingsDir: "src/custom-typings",
-  tscOptions: <any> {
+  tscOptions: {
+    declaration: true,
     skipLibCheck: true,
   },
   typedoc: {
@@ -59,27 +61,24 @@ const lib: buildTools.LibTarget = {
   },
 };
 
-const test: buildTools.MochaTarget = {
+const test: MochaTarget = {
   project,
   name: "test",
   srcDir: "src",
   scripts: ["test/**/*.ts", "lib/**/*.ts"],
-  outModules: buildTools.OutModules.Js,
   customTypingsDir: "src/custom-typings",
-  tscOptions: <any> {
+  tscOptions: {
     skipLibCheck: true,
-    allowSyntheticDefaultImports: true,
-    esModuleInterop: true,
   },
-  copy: [{files: ["test/scrapping/**/*.html"]}],
   clean: {
     dirs: ["build/test"],
   },
 };
 
-const libTasks: any = buildTools.registerLibTasks(gulp, lib);
-buildTools.registerMochaTasks(gulp, test);
+const libTasks: any = registerLibTasks(gulp, lib);
+registerMochaTasks(gulp, test);
 buildTools.projectTasks.registerAll(gulp, project);
 
 gulp.task("all:tsconfig.json", gulp.parallel("lib:tsconfig.json", "test:tsconfig.json"));
 gulp.task("dist", libTasks.dist);
+gulp.task("default", libTasks.dist);